Made pr review edits and moved into the ramp folder
