+++ src/main/java/frc/robot/Robot.java
@@ -10,9 +10,9 @@ import edu.wpi.first.wpilibj.TimedRobot;
 import edu.wpi.first.wpilibj.Timer;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.CommandScheduler;
-import frc.robot.commands.ResetRamp;
 import frc.robot.commands.drivetrain.ResetGyro;
 import frc.robot.commands.drivetrain.WheelAlign;
+import frc.robot.commands.ramp.ResetRamp;
 import frc.robot.constants.Constants;
 import frc.robot.utils.diag.Diagnostics;
 import frc.robot.utils.logging.Logger;
